refactor(filterbar): rename for avoid conflict on autoimport component
