+++ src/components/filterbar/pinned/PinnedToggle.vue
